h-c: add headings for each type of comparison
Without headings, it's non-obvious which sections correspond to which
comparisons in the output file.  This information was printed to
standard output, but not to the output file.  So I added this
information to the output file as well.

My knowledge of Perl is limited so I just used system("/bin/echo ..");
for now.  It would be good to fix that later.

I've added a copyright notice for myself, since this change is
probably significant enough to be copyrightable.