h-c: infinite context for file MD5 sum comparison
Since C&CS checkers are often interested in the list of files that
have stayed the same between a firmware release and the result of
compiling the C&CS for that firmware (ie. to see if there are binaries
which have not changed, which is unlikely because each compiler
configuration outputs slightly different binaries), add "infinite"
lines of context in the diff command for the MD5 checksum comparison.
This will include all files that have not changed (in addition to the
files that have changed).

I didn't immediately see a flag for infinite lines of context in the
diff manpage so I'm using a big number for now (2,000,000,000).