Working version of forking central commit.
Note that due to the fact that Git::Repository calls System::Command,
you have to be careful about when you install the child signal handler.

You only want it in the parent, after the parent has done its operations
with Git commands, and note that the child must also disable the handler.